move mask gen funciton into ascend backend
